Refactor transformation of fetch data to an array in the store
